+++ distro/config.go
@@ -297,20 +297,3 @@ func WithLogger(l logr.Logger) Option {
 		c.Logger = l
 	})
 }
-
-// WithSpanLimits configures the span limits used by the distro.
-//
-// The OpenTelemetry environment limit variables
-// (OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT, OTEL_ATTRIBUTE_COUNT_LIMIT,
-// OTEL_SPAN_ATTRIBUTE_VALUE_LENGTH_LIMIT, OTEL_SPAN_ATTRIBUTE_COUNT_LIMIT,
-// OTEL_SPAN_EVENT_COUNT_LIMIT, OTEL_SPAN_LINK_COUNT_LIMIT,
-// OTEL_EVENT_ATTRIBUTE_COUNT_LIMIT, OTEL_LINK_ATTRIBUTE_COUNT_LIMIT) are used
-// if this option is not provided.
-//
-// By default, the link count is limited to 1000, the attribute value length
-// is limited to 12000, and all other limts are set to be unlimited.
-func WithSpanLimits(limits trace.SpanLimits) Option {
-	return optionFunc(func(c *config) {
-		c.SpanLimits = &limits
-	})
-}