+++ CHANGELOG.md
@@ -22,6 +22,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Add the `WithLogger` option to `github.com/signalfx/splunk-otel-go/distro`
   along with parsing of the `OTEL_LOG_LEVEL` environment variable to set the
   logging level of the default logger used. (#336)
+- Add the `WithSpanLimits` option to
+  `github.com/signalfx/splunk-otel-go/distro` to set the span limits used.
+  (#723)
 
 ### Changed