+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+## [0.8.0] - 2022-04-01
+
 ### Added
 
 - Add the `WithPropagator` option to
@@ -211,7 +213,8 @@ an impedance mismatch with this duplicate batching.
 - Add [`splunkhttp`](./instrumentation/net/http/splunkhttp) module providing
   additional Splunk specific instrumentation for `net/http`.
 
-[Unreleased]: https://github.com/signalfx/splunk-otel-go/compare/v0.7.0...HEAD
+[Unreleased]: https://github.com/signalfx/splunk-otel-go/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.8.0
 [0.7.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.7.0
 [0.6.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.6.0
 [0.4.0]: https://github.com/signalfx/splunk-otel-go/releases/tag/v0.4.0